隔离所有组件
