+++ components/dialog/dialog.wxml
