+++ components/qrcode/qr.js/index.js
