+++ components/qrcode/qr.js/lib/8BitByte.js
