+++ components/qrcode/qr.js/lib/BitBuffer.js
