+++ components/qrcode/qr.js/lib/ErrorCorrectLevel.js
