+++ components/qrcode/qr.js/lib/LICENSE-qrcode
