+++ components/qrcode/qr.js/lib/Polynomial.js
