+++ components/qrcode/qr.js/lib/QRCode.js
