+++ components/qrcode/qr.js/lib/RSBlock.js
