+++ components/qrcode/qr.js/lib/math.js
