+++ components/qrcode/qr.js/lib/mode.js
