+++ components/qrcode/qr.js/lib/util.js
