+++ components/rater/rater.wxml
