+++ pages/index/index.js
@@ -2,52 +2,52 @@ Page({
 	data: {
 		components: [
 			{
-				title: 'backdrop', 
+				title: 'Backdrop', 
 				url: '/pages/backdrop/index', 
 				icon: '../../assets/images/iconfont-backdrop.png', 
 			},
 			{
-				title: 'dialog', 
+				title: 'Dialog', 
 				url: '/pages/dialog/index', 
 				icon: '../../assets/images/iconfont-dialog.png', 
 			},
 			{
-				title: 'toast', 
-				url: '/pages/toast/index', 
-				icon: '../../assets/images/iconfont-toast.png', 
-			},
-			{
-				title: 'loading', 
-				url: '/pages/loading/index', 
-				icon: '../../assets/images/iconfont-loading.png', 
-			},
-			{
-				title: 'rater', 
-				url: '/pages/rater/index', 
-				icon: '../../assets/images/iconfont-rater.png', 
-			},
-			{
-				title: 'picker', 
-				url: '/pages/picker/index', 
-				icon: '../../assets/images/iconfont-picker.png', 
-			},
-			{
-				title: 'toptips', 
-				url: '/pages/toptips/index', 
-				icon: '../../assets/images/iconfont-toptips.png', 
-			},
-			{
-				title: 'qrcode', 
-				url: '/pages/qrcode/index', 
-				icon: '../../assets/images/iconfont-qrcode.png', 
-			},
-			{
-				title: 'gallery', 
+				title: 'Gallery', 
 				url: '/pages/gallery/index', 
 				icon: '../../assets/images/iconfont-gallery.png', 
 			},
 			{
-				title: 'xnumber', 
+				title: 'Loading', 
+				url: '/pages/loading/index', 
+				icon: '../../assets/images/iconfont-loading.png', 
+			},
+			{
+				title: 'Picker', 
+				url: '/pages/picker/index', 
+				icon: '../../assets/images/iconfont-picker.png', 
+			},
+			{
+				title: 'Qrcode', 
+				url: '/pages/qrcode/index', 
+				icon: '../../assets/images/iconfont-qrcode.png', 
+			},
+			{
+				title: 'Rater', 
+				url: '/pages/rater/index', 
+				icon: '../../assets/images/iconfont-rater.png', 
+			},
+			{
+				title: 'Toast', 
+				url: '/pages/toast/index', 
+				icon: '../../assets/images/iconfont-toast.png', 
+			},
+			{
+				title: 'Toptips', 
+				url: '/pages/toptips/index', 
+				icon: '../../assets/images/iconfont-toptips.png', 
+			},
+			{
+				title: 'Xnumber', 
 				url: '/pages/xnumber/index', 
 				icon: '../../assets/images/iconfont-xnumber.png', 
 			},
