+++ pages/rater/index.wxml
@@ -1,4 +1,4 @@
-<import src="../../components/rater.wxml"/>
+<import src="../../components/rater/rater.wxml"/>
 
 <view class="page">
     <view class="page__hd">
