+++ pages/xnumber/index.wxml
@@ -1,4 +1,4 @@
-<import src="../../components/xnumber.wxml"/>
+<import src="../../components/xnumber/xnumber.wxml"/>
 
 <view class="page">
     <view class="page__hd">
